Refactored Product form to be submitted in the main page component. fixed error that caused the page to not rout back to products page
